+++ app/javascript/packs/application.js
@@ -1,14 +1,6 @@
 import "bootstrap";
 import { initMapbox } from "../plugins/init_mapbox.js";
+import { getPosition } from "../plugins/mapbox_position.js";
 
-const getPosition = (event) => {
-  navigator.geolocation.getCurrentPosition((data) => {
-    const lat = data.coords.latitude;
-    const long = data.coords.longitude;
-    const coordinates = [long, lat];
-    console.log(coordinates);
-    initMapbox(coordinates);
-  });
-}
-
+initMapbox();
 getPosition();
